fix: git date: echo => sed
This fixes an issue for which the "-e" option of the "echo" command in
the "out/gitdate.tex" target would end up in the .tex file.
The issue was fixed by using a file src/gitdate.tex with a placeholder
that gets replaced using sed.

Updated README.md with instructions to build documents on Ubuntu Linux.